backend/requirements.txt
Flask==2.3.2
Flask-CORS==3.0.10
pymongo==4.3.3
requests==2.31.0
networkx==3.1
matplotlib==3.7.2
python-dotenv==1.0.0
